Move ConvertEmissiveUnitFunctor to public interface of Atom_Feature_Common
