+++ Gems/Atom/Feature/Common/Code/Include/Atom/Feature/Material/ConvertEmissiveUnitFunctor.h
