+++ Gems/Atom/Feature/Common/Code/atom_feature_common_files.cmake
@@ -114,8 +114,6 @@ set(FILES
     Source/ImGui/ImGuiSystemComponent.h
     Source/ImageBasedLights/ImageBasedLightFeatureProcessor.cpp
     Source/LookupTable/LookupTableAsset.cpp
-    Source/Material/ConvertEmissiveUnitFunctor.cpp
-    Source/Material/ConvertEmissiveUnitFunctor.h
     Source/Material/SubsurfaceTransmissionParameterFunctor.cpp
     Source/Material/SubsurfaceTransmissionParameterFunctor.h
     Source/Material/Transform2DFunctor.cpp
